+++ packages/databyss-ui/editor/slate/markup.js
@@ -1,29 +1,6 @@
 import cloneDeep from 'clone-deep'
 import { Editor, Value } from 'slate'
-import ObjectId from 'bson-objectid'
 import { isAtomicInlineType } from './page/reducer'
-import Html from 'slate-html-serializer'
-
-const MARK_TAG = {
-  em: 'italic',
-  strong: 'bold',
-}
-
-const rules = [
-  {
-    deserialize(el, next) {
-      const mark = MARK_TAG[el.tagName.toLowerCase()]
-      if (!mark) {
-        return
-      }
-      return {
-        object: 'mark',
-        type: mark,
-        nodes: next(el.childNodes),
-      }
-    },
-  },
-]
 
 export const getRangesFromBlock = block => {
   const { nodes } = block
@@ -70,43 +47,6 @@ export const slateToState = (slate, _id) => {
     ranges,
   }
   return { [slate.key]: response }
-}
-
-/*
-takes a node list and deserializes them to return a list with refId, _id, text, and ranges
-*/
-export const blocksToState = nodes => {
-  const _blocks = nodes.map(block => blockToState(block)).toJS()
-  return _blocks
-}
-
-export const blockToState = block => {
-  // refID is required in the block data
-  // refId is used to look up ranges and text in state
-  let refId = block.data ? block.data.get('refId') : null
-  let _textFields
-  let _block = block
-
-  if (isAtomicInlineType(block.type)) {
-    // deserializes the html text to return ranges and marks
-    _block = new Html({ rules }).deserialize(block.text).anchorBlock
-  } else {
-    // if not atomic, generate new refId
-    refId = ObjectId().toHexString()
-  }
-  _textFields = getRangesFromBlock(_block.toJSON())
-
-  const text = _textFields.text
-  const ranges = _textFields.ranges
-
-  const response = {
-    text,
-    type: block.type,
-    ranges,
-    refId,
-    _id: block.key,
-  }
-  return { [block.key]: response }
 }
 
 export const stateToSlateMarkup = state => {
