+++ packages/databyss-ui/editor/slate/page/ContentEditable.js
@@ -1,7 +1,6 @@
 import React, { useRef, useEffect, forwardRef } from 'react'
 import { Value } from 'slate'
 import { Editor, getEventTransfer } from 'slate-react'
-import ObjectId from 'bson-objectid'
 import forkRef from '@databyss-org/ui/lib/forkRef'
 import Bugsnag from '@databyss-org/services/lib/bugsnag'
 import { getRawHtmlForBlock } from './../../state/page/reducer'
@@ -22,7 +21,7 @@ import {
   newEditor,
 } from './../slateUtils'
 
-import { blocksToState } from './../markup'
+import { blocksToState, getFragFromText } from './../clipboard'
 
 const schema = {
   inlines: {
@@ -399,15 +398,12 @@ const SlateContentEditable = forwardRef(
 
     const onPaste = (event, editor, next) => {
       // TODO: if html convert to ranges
-
       const { value } = editor
       const transfer = getEventTransfer(event)
       const { fragment, type } = transfer
       let _frag = fragment
       // get anchor block from slate,
       const anchorKey = value.anchorBlock.key
-      console.log('transfer', transfer)
-      // TODO: CHECK FOR TEXT CARRYIGN CARRIGE RETURNS
       if (type === 'fragment') {
         if (_frag.nodes.size > 1) {
           // trim first node if empty
@@ -428,46 +424,10 @@ const SlateContentEditable = forwardRef(
         onPasteAction(anchorKey, _blockList, _frag, editor)
         return event.preventDefault()
       }
-      // if plaintext or html
-
-      // creates list of new blocks with refId and _id
-
-      // create a list split by carriage returns
-      const _textList = transfer.text.split(/\r?\n/)
-      // creates a slate editor to compose a fragment
-      let _editor = newEditor()
-      const _blockList = _textList.map(t => {
-        const _refId = ObjectId().toHexString()
-        const _key = ObjectId().toHexString()
-        const _block = {
-          object: 'block',
-          type: 'ENTRY',
-          data: { refId: _refId },
-          key: _key,
-          nodes: [
-            {
-              object: 'text',
-              text: t,
-            },
-          ],
-        }
-        _editor.insertBlock(_block)
-        return {
-          [_key]: {
-            _id: _key,
-            refId: _refId,
-            text: t,
-            type: 'ENTRY',
-            ranges: [],
-          },
-        }
-      })
-      // removes first node in fragment
-      // this node is empty by default
-      _frag = _editor.value.document.removeNode(
-        _editor.value.document.nodes.get(0).key
-      )
-
+      // if plaintext or html is pasted
+      const _textData = getFragFromText(transfer.text)
+      const _blockList = _textData._blockList
+      _frag = _textData._frag
       onPasteAction(anchorKey, _blockList, _frag, editor)
       return event.preventDefault()
     }
